requirements.txt
streamlit==1.45.1
PyPDF2==3.0.1
langchain==0.3.25
langchain-google-genai==0.1.3
langchain-text-splitters==0.0.2
faiss-cpu==1.8.0
python-dotenv==1.0.1
google-generativeai==0.3.0